+++ packages/default-layout/__tests__/fitPageWidthOpenSidebar.e2e.ts
@@ -1,0 +1,30 @@
+import 'expect-puppeteer';
+
+test('Keep to fit page width when click sidebar', async () => {
+    await page.goto('http://localhost:3000/default-layout-fit-page-with');
+    await page.setViewport({
+        width: 1920,
+        height: 1080,
+    });
+    await page.evaluate(() => document.querySelector('[data-testid="core__viewer"]').scrollIntoView());
+
+    // Wait until the first page is rendered
+    await page.waitForSelector('[data-testid="core__text-layer-0"]', { visible: true });
+
+    // Click the `Thumbnail` tab in the sidebar
+    const thumbnailTab = await page.waitForSelector('button[aria-label="Thumbnail"]');
+    await thumbnailTab.click();
+
+    // To make sure the thumbnails are rendered
+    await page.waitForSelector('[data-testid="thumbnail__list"]');
+
+    // Wait until the thumbnail of second page is rendered
+    await page.waitForFunction(() => document.querySelector('[aria-label="Thumbnail of page 2"]'));
+
+    // Wait until the 4th page is rendered
+    await page.waitForSelector('[data-testid="core__text-layer-3"]');
+
+    const currentZoomLabel = await page.waitForSelector('[data-testid="zoom__popover-target-scale"]');
+    const currentZoom = await currentZoomLabel.evaluate((ele) => ele.textContent);
+    expect(currentZoom).toEqual('115%');
+});